requirements.txt
torch>=1.9.0
transformers>=4.21.0
gradio>=3.0.0
Pillow>=8.3.0
requests>=2.25.0
numpy>=1.21.0
flask>=2.0.0
flask-cors>=3.0.0
gunicorn>=20.1.0
streamlit>=1.0.0
